integration with balkis commit

+++ Pidev-Sprint1-ejb/src/main/java/Interfaces/OperationInterface.java
@@ -10,10 +10,10 @@ import javax.ejb.Remote;
 @Remote
 public interface OperationInterface {
 	public void addOperation (Operation op);
-	public void verser(double montant,int Bank_Acount_id,int User_ID);
-	public void retirer(double montant,int Bank_Acount_id,int User_ID);
+	public void verser(float montant,int Bank_Acount_id,int User_ID);
+	public void retirer(float montant,int Bank_Acount_id,int User_ID);
 	public BankAccount consultercompte(int Bank_Acount_id);
 	public List<Operation> consulterOperations(int Bank_Acount_id);
-	public void UpdateBalance(int id, double balance);
+	public void UpdateBalance(int id, float balance);
 }
